+++ ssbhesabfa.php
@@ -173,7 +173,7 @@ class Ssbhesabfa extends Module
     {
         $this->name = 'ssbhesabfa';
         $this->tab = 'billing_invoicing';
-        $this->version = '2.3.26';
+        $this->version = '2.3.27';
         $this->author = 'Saeed Sattar Beglou';
         $this->need_instance = 0;